More efficient and move things around
