+++ internal/pkg/describe/backend_service_test.go
@@ -311,15 +311,10 @@ func TestBackendServiceDescriber_Describe(t *testing.T) {
 						},
 					},
 					ServiceDiscovery: serviceDiscoveries{
-						{
-							Environment: []string{"test"},
-							Endpoint:    "jobs.test.phonetool.local:5000",
-						},
-						{
-							Environment: []string{"prod"},
-							Endpoint:    "jobs.prod.phonetool.local:5000",
-						},
+						"jobs.test.phonetool.local:5000": []string{"test"},
+						"jobs.prod.phonetool.local:5000": []string{"prod"},
 					},
+					ServiceConnect: serviceConnects{},
 					Variables: []*containerEnvVar{
 						{
 							envVar: &envVar{
@@ -460,16 +455,10 @@ func TestBackendServiceDescriber_Describe(t *testing.T) {
 						},
 					},
 					ServiceDiscovery: serviceDiscoveries{
-						{
-							Environment: []string{"test"},
-							Endpoint:    "jobs.test.phonetool.local:5000",
-						},
+						"jobs.test.phonetool.local:5000": []string{"test"},
 					},
 					ServiceConnect: serviceConnects{
-						{
-							Environment: []string{"test"},
-							Endpoint:    "jobs",
-						},
+						"jobs": []string{"test"},
 					},
 					Variables: []*containerEnvVar{
 						{
@@ -570,8 +559,8 @@ Internal Service Endpoint
 
   Endpoint                              Environment  Type
   --------                              -----------  ----
-  http://my-svc.test.my-app.local:5000  test         Service Discovery
   http://my-svc.prod.my-app.local:5000  prod         Service Discovery
+  http://my-svc.test.my-app.local:5000  test         Service Discovery
 
 Variables
 
@@ -595,7 +584,7 @@ Resources
   prod
     AWS::EC2::SecurityGroupIngress  ContainerSecurityGroupIngressFromPublicALB
 `,
-			wantedJSONString: "{\"service\":\"my-svc\",\"type\":\"Backend Service\",\"application\":\"my-app\",\"configurations\":[{\"environment\":\"test\",\"port\":\"80\",\"cpu\":\"256\",\"memory\":\"512\",\"platform\":\"LINUX/X86_64\",\"tasks\":\"1\"},{\"environment\":\"prod\",\"port\":\"5000\",\"cpu\":\"512\",\"memory\":\"1024\",\"platform\":\"LINUX/ARM64\",\"tasks\":\"3\"}],\"routes\":null,\"serviceDiscovery\":[{\"environment\":[\"test\"],\"endpoint\":\"http://my-svc.test.my-app.local:5000\"},{\"environment\":[\"prod\"],\"endpoint\":\"http://my-svc.prod.my-app.local:5000\"}],\"variables\":[{\"environment\":\"prod\",\"name\":\"COPILOT_ENVIRONMENT_NAME\",\"value\":\"prod\",\"container\":\"container\"},{\"environment\":\"test\",\"name\":\"COPILOT_ENVIRONMENT_NAME\",\"value\":\"test\",\"container\":\"container\"}],\"secrets\":[{\"name\":\"GITHUB_WEBHOOK_SECRET\",\"container\":\"container\",\"environment\":\"test\",\"valueFrom\":\"GH_WEBHOOK_SECRET\"},{\"name\":\"SOME_OTHER_SECRET\",\"container\":\"container\",\"environment\":\"prod\",\"valueFrom\":\"SHHHHH\"}],\"resources\":{\"prod\":[{\"type\":\"AWS::EC2::SecurityGroupIngress\",\"physicalID\":\"ContainerSecurityGroupIngressFromPublicALB\"}],\"test\":[{\"type\":\"AWS::EC2::SecurityGroup\",\"physicalID\":\"sg-0758ed6b233743530\"}]}}\n",
+			wantedJSONString: "{\"service\":\"my-svc\",\"type\":\"Backend Service\",\"application\":\"my-app\",\"configurations\":[{\"environment\":\"test\",\"port\":\"80\",\"cpu\":\"256\",\"memory\":\"512\",\"platform\":\"LINUX/X86_64\",\"tasks\":\"1\"},{\"environment\":\"prod\",\"port\":\"5000\",\"cpu\":\"512\",\"memory\":\"1024\",\"platform\":\"LINUX/ARM64\",\"tasks\":\"3\"}],\"routes\":null,\"serviceDiscovery\":[{\"environment\":[\"prod\"],\"endpoint\":\"http://my-svc.prod.my-app.local:5000\"},{\"environment\":[\"test\"],\"endpoint\":\"http://my-svc.test.my-app.local:5000\"}],\"variables\":[{\"environment\":\"prod\",\"name\":\"COPILOT_ENVIRONMENT_NAME\",\"value\":\"prod\",\"container\":\"container\"},{\"environment\":\"test\",\"name\":\"COPILOT_ENVIRONMENT_NAME\",\"value\":\"test\",\"container\":\"container\"}],\"secrets\":[{\"name\":\"GITHUB_WEBHOOK_SECRET\",\"container\":\"container\",\"environment\":\"test\",\"valueFrom\":\"GH_WEBHOOK_SECRET\"},{\"name\":\"SOME_OTHER_SECRET\",\"container\":\"container\",\"environment\":\"prod\",\"valueFrom\":\"SHHHHH\"}],\"resources\":{\"prod\":[{\"type\":\"AWS::EC2::SecurityGroupIngress\",\"physicalID\":\"ContainerSecurityGroupIngressFromPublicALB\"}],\"test\":[{\"type\":\"AWS::EC2::SecurityGroup\",\"physicalID\":\"sg-0758ed6b233743530\"}]}}\n",
 		},
 	}
 
@@ -656,14 +645,8 @@ Resources
 				},
 			}
 			sds := serviceDiscoveries{
-				{
-					Environment: []string{"test"},
-					Endpoint:    "http://my-svc.test.my-app.local:5000",
-				},
-				{
-					Environment: []string{"prod"},
-					Endpoint:    "http://my-svc.prod.my-app.local:5000",
-				},
+				"http://my-svc.test.my-app.local:5000": []string{"test"},
+				"http://my-svc.prod.my-app.local:5000": []string{"prod"},
 			}
 			resources := map[string][]*stack.Resource{
 				"test": {
